修改主键泛型为Int类型

+++ src/main/java/com/lord/advertis/dao/CustomersDao.java
@@ -3,6 +3,6 @@ package com.lord.advertis.dao;
 import com.lord.advertis.base.BaseMapper;
 import com.lord.advertis.domain.Customers;
 
-public interface CustomersDao extends BaseMapper<Customers, Long>{
+public interface CustomersDao extends BaseMapper<Customers, Integer>{
 
 }
